+++ src/wallets/web3-provider/methods/utils.js
@@ -1,0 +1,25 @@
+import { formatters } from 'web3-core-helpers';
+const getSanitizedTx = tx => {
+  return new Promise((resolve, reject) => {
+    if (!tx.gas && !tx.gasLimit && !tx.chainId)
+      return reject(new Error('"gas" or "chainId" is missing'));
+    if (tx.nonce < 0 || tx.gas < 0 || tx.gasPrice < 0 || tx.chainId < 0)
+      return reject(
+        new Error('Gas, gasPrice, nonce or chainId is lower than 0')
+      );
+
+    try {
+      tx = formatters.inputCallFormatter(tx);
+      const transaction = tx;
+      transaction.to = tx.to || '0x';
+      transaction.data = tx.data || '0x';
+      transaction.value = tx.value || '0x';
+      transaction.chainId = tx.chainId;
+      resolve(transaction);
+    } catch (e) {
+      reject(e);
+    }
+  });
+};
+
+export { getSanitizedTx };
